Add files via upload
Fill in the appropriate sections:

Section 2 Requirements: Add in or amend the requirements section of what you are working on. 

Section 3: Design & Components - add in what you are going to actually build to satisfy the requirements. 

Section 4: Add in specifically what you are working on. 

Section 6: add in team information and git info